+++ pgcrypto_fields/fields.py
@@ -3,35 +3,16 @@ from __future__ import unicode_literals
 import datetime
 
 from django.db import models
+from django.db.utils import six
 
 from pgcrypto_fields import (
-    DIGEST_SQL,
-    HMAC_SQL,
     INTEGER_PGP_PUB_ENCRYPT_SQL,
-    INTEGER_PGP_SYM_ENCRYPT_SQL,
     PGP_PUB_ENCRYPT_SQL,
-    PGP_SYM_ENCRYPT_SQL,
 )
-from pgcrypto_fields.lookups import DigestLookup, HMACLookup
 from pgcrypto_fields.mixins import (
     EmailPGPPublicKeyFieldMixin,
-    EmailPGPSymmetricKeyFieldMixin,
-    HashMixin,
     PGPPublicKeyFieldMixin,
-    PGPSymmetricKeyFieldMixin,
 )
-
-
-class TextDigestField(HashMixin, models.TextField):
-    """Text digest field for postgres."""
-    encrypt_sql = DIGEST_SQL
-TextDigestField.register_lookup(DigestLookup)
-
-
-class TextHMACField(HashMixin, models.TextField):
-    """Text HMAC field for postgres."""
-    encrypt_sql = HMAC_SQL
-TextHMACField.register_lookup(HMACLookup)
 
 
 class EmailPGPPublicKeyField(EmailPGPPublicKeyFieldMixin, models.EmailField):
@@ -42,6 +23,12 @@ class EmailPGPPublicKeyField(EmailPGPPublicKeyFieldMixin, models.EmailField):
 class IntegerPGPPublicKeyField(PGPPublicKeyFieldMixin, models.IntegerField):
     """Integer PGP public key encrypted field."""
     encrypt_sql = INTEGER_PGP_PUB_ENCRYPT_SQL
+
+    @classmethod
+    def _parse_decrypted_value(cls, value):
+        if not isinstance(value, six.string_types):
+            return value
+        return int(value)
 
 
 class TextPGPPublicKeyField(PGPPublicKeyFieldMixin, models.TextField):
@@ -56,6 +43,9 @@ class DatePGPPublicKeyField(PGPPublicKeyFieldMixin, models.DateField):
 
     @classmethod
     def _parse_decrypted_value(cls, value):
+        if not isinstance(value, six.string_types):
+            return value
+
         if value is None:
             return None
         return datetime.datetime.strptime(value, "%Y-%m-%d").date()
@@ -85,6 +75,9 @@ class NullBooleanPGPPublicKeyField(PGPPublicKeyFieldMixin, models.NullBooleanFie
 
     @classmethod
     def _parse_decrypted_value(cls, value):
+        if not isinstance(value, six.string_types):
+            return value
+
         if value == 'True':
             value = True
         elif value == 'False':
@@ -95,18 +88,3 @@ class NullBooleanPGPPublicKeyField(PGPPublicKeyFieldMixin, models.NullBooleanFie
                 'Value: %s, type: %s' % (value, type(value))
             )
         return value
-
-
-class EmailPGPSymmetricKeyField(EmailPGPSymmetricKeyFieldMixin, models.EmailField):
-    """Email PGP symmetric key encrypted field."""
-    encrypt_sql = PGP_SYM_ENCRYPT_SQL
-
-
-class IntegerPGPSymmetricKeyField(PGPSymmetricKeyFieldMixin, models.IntegerField):
-    """Integer PGP symmetric key encrypted field."""
-    encrypt_sql = INTEGER_PGP_SYM_ENCRYPT_SQL
-
-
-class TextPGPSymmetricKeyField(PGPSymmetricKeyFieldMixin, models.TextField):
-    """Text PGP symmetric key encrypted field for postgres."""
-    encrypt_sql = PGP_SYM_ENCRYPT_SQL