+++ tests/test_fields.py
@@ -9,24 +9,12 @@ from .factories import EncryptedModelFactory
 from .models import EncryptedModel
 
 
-KEYED_FIELDS = (fields.TextDigestField, fields.TextHMACField)
-EMAIL_PGP_FIELDS = (fields.EmailPGPPublicKeyField, fields.EmailPGPSymmetricKeyField)
+EMAIL_PGP_FIELDS = (fields.EmailPGPPublicKeyField,)
 PGP_FIELDS = EMAIL_PGP_FIELDS + (
     fields.IntegerPGPPublicKeyField,
-    fields.IntegerPGPSymmetricKeyField,
     fields.TextPGPPublicKeyField,
-    fields.TextPGPSymmetricKeyField,
     fields.DatePGPPublicKeyField,
 )
-
-
-class TestTextFieldHash(TestCase):
-    """Test hash fields behave properly."""
-    def test_get_placeholder(self):
-        """Assert `get_placeholder` hash value only once."""
-        for field in KEYED_FIELDS:
-            placeholder = field().get_placeholder('\\x')
-            self.assertEqual(placeholder, '%s')
 
 
 class TestPGPMixin(TestCase):
@@ -65,14 +53,9 @@ class TestEncryptedTextFieldModel(TestCase):
         fields = self.model._meta.get_all_field_names()
         expected = (
             'id',
-            'digest_field',
-            'hmac_field',
             'email_pgp_pub_field',
             'integer_pgp_pub_field',
             'pgp_pub_field',
-            'email_pgp_sym_field',
-            'integer_pgp_sym_field',
-            'pgp_sym_field',
             'pgp_pub_date_field',
             'pgp_pub_null_boolean_field',
         )
@@ -83,25 +66,15 @@ class TestEncryptedTextFieldModel(TestCase):
         EncryptedModelFactory.create()
 
         instance = self.model.objects.get()
-        self.assertIsInstance(instance.digest_field, unicode)
-        self.assertIsInstance(instance.hmac_field, unicode)
 
         self.assertIsInstance(instance.email_pgp_pub_field, unicode)
         self.assertIsInstance(instance.integer_pgp_pub_field, int)
         self.assertIsInstance(instance.pgp_pub_field, unicode)
 
-        self.assertIsInstance(instance.email_pgp_sym_field, unicode)
-        self.assertIsInstance(instance.integer_pgp_sym_field, int)
-        self.assertIsInstance(instance.pgp_sym_field, unicode)
-
     def test_fields_descriptor_is_not_instance(self):
         """`EncryptedProxyField` instance returns itself when accessed from the model."""
         self.assertIsInstance(
             self.model.pgp_pub_field,
-            proxy.EncryptedProxyField,
-        )
-        self.assertIsInstance(
-            self.model.pgp_sym_field,
             proxy.EncryptedProxyField,
         )
 
@@ -112,14 +85,13 @@ class TestEncryptedTextFieldModel(TestCase):
 
         instance = self.model.objects.get()
 
-        with self.assertNumQueries(1):
+        with self.assertNumQueries(0):
             instance.pgp_pub_field
 
     def test_value_pgp_pub(self):
         """Assert we can get back the decrypted value."""
         expected = 'bonjour'
         EncryptedModelFactory.create(pgp_pub_field=expected)
-
         instance = self.model.objects.get()
         value = instance.pgp_pub_field
         self.assertEqual(value, expected)
@@ -164,16 +136,6 @@ class TestEncryptedTextFieldModel(TestCase):
 
         self.assertEqual(value, expected)
 
-    def test_value_pgp_sym(self):
-        """Assert we can get back the decrypted value."""
-        expected = 'bonjour'
-        EncryptedModelFactory.create(pgp_sym_field=expected)
-
-        instance = self.model.objects.get()
-        value = instance.pgp_sym_field
-
-        self.assertEqual(value, expected)
-
     def test_instance_not_saved(self):
         """Assert not saved instance return the value to be encrypted."""
         expected = 'bonjour'
@@ -186,16 +148,13 @@ class TestEncryptedTextFieldModel(TestCase):
         expected = 'bonjour'
         EncryptedModelFactory.create(
             pgp_pub_field=expected,
-            pgp_sym_field=expected,
         )
 
         queryset = self.model.objects.annotate(
             aggregates.PGPPublicKeyAggregate('pgp_pub_field'),
-            aggregates.PGPSymmetricKeyAggregate('pgp_sym_field'),
         )
         instance = queryset.get()
         self.assertEqual(instance.pgp_pub_field__decrypted, expected)
-        self.assertEqual(instance.pgp_sym_field__decrypted, expected)
 
     def test_decrypt_filter(self):
         """Assert we can get filter the decrypted value."""
@@ -210,50 +169,6 @@ class TestEncryptedTextFieldModel(TestCase):
         instance = queryset.filter(pgp_pub_field__decrypted=expected).first()
         self.assertEqual(instance.pgp_pub_field__decrypted, expected)
 
-    def test_digest_lookup(self):
-        """Assert we can filter a digest value."""
-        value = 'bonjour'
-        expected = EncryptedModelFactory.create(digest_field=value)
-        EncryptedModelFactory.create()
-
-        queryset = EncryptedModel.objects.filter(digest_field__hash_of=value)
-
-        self.assertItemsEqual(queryset, [expected])
-
-    def test_hmac_lookup(self):
-        """Assert we can filter a digest value."""
-        value = 'bonjour'
-        expected = EncryptedModelFactory.create(hmac_field=value)
-        EncryptedModelFactory.create()
-
-        queryset = EncryptedModel.objects.filter(hmac_field__hash_of=value)
-        self.assertItemsEqual(queryset, [expected])
-
-    def test_default_lookup(self):
-        """Assert default lookup can be called."""
-        queryset = EncryptedModel.objects.filter(hmac_field__isnull=True)
-        self.assertFalse(queryset)
-
-    def test_update_attribute_digest_field(self):
-        """Assert digest field can be updated through its attribute on the model."""
-        expected = 'bonjour'
-        instance = EncryptedModelFactory.create()
-        instance.digest_field = expected
-        instance.save()
-
-        updated_instance = self.model.objects.filter(digest_field__hash_of=expected)
-        self.assertEqual(updated_instance.first(), instance)
-
-    def test_update_attribute_hmac_field(self):
-        """Assert hmac field can be updated through its attribute on the model."""
-        expected = 'bonjour'
-        instance = EncryptedModelFactory.create()
-        instance.hmac_field = expected
-        instance.save()
-
-        updated_instance = self.model.objects.filter(hmac_field__hash_of=expected)
-        self.assertEqual(updated_instance.first(), instance)
-
     def test_update_attribute_pgp_pub_field(self):
         """Assert pgp field can be updated through its attribute on the model."""
         expected = 'bonjour'
@@ -264,16 +179,6 @@ class TestEncryptedTextFieldModel(TestCase):
         updated_instance = self.model.objects.get()
         self.assertEqual(updated_instance.pgp_pub_field, expected)
 
-    def test_update_attribute_pgp_sym_field(self):
-        """Assert pgp field can be updated through its attribute on the model."""
-        expected = 'bonjour'
-        instance = EncryptedModelFactory.create()
-        instance.pgp_sym_field = expected
-        instance.save()
-
-        updated_instance = self.model.objects.get()
-        self.assertEqual(updated_instance.pgp_sym_field, expected)
-
     def test_update_one_attribute(self):
         """Assert value are not overriden when updating one attribute."""
         expected = 'initial value'
@@ -281,22 +186,12 @@ class TestEncryptedTextFieldModel(TestCase):
 
         instance = EncryptedModelFactory.create(
             pgp_pub_field=expected,
-            pgp_sym_field=expected,
-            digest_field=expected,
-            hmac_field=expected,
         )
         instance.pgp_sym_field = new_value
         instance.save()
 
         updated_instance = self.model.objects.get()
         self.assertEqual(updated_instance.pgp_pub_field, expected)
-        self.assertEqual(updated_instance.pgp_sym_field, new_value)
-
-        updated_instance = self.model.objects.filter(
-            digest_field__hash_of=expected,
-            hmac_field__hash_of=expected,
-        )
-        self.assertEqual(updated_instance.first(), instance)
 
     def test_pgp_public_key_negative_number(self):
         """Assert negative value is saved with an `IntegerPGPPublicKeyField` field."""
@@ -304,13 +199,6 @@ class TestEncryptedTextFieldModel(TestCase):
         instance = EncryptedModelFactory.create(integer_pgp_pub_field=expected)
 
         self.assertEqual(instance.integer_pgp_pub_field, expected)
-
-    def test_pgp_symmetric_key_negative_number(self):
-        """Assert negative value is saved with an `IntegerPGPSymmetricKeyField` field."""
-        expected = -1
-        instance = EncryptedModelFactory.create(integer_pgp_sym_field=expected)
-
-        self.assertEqual(instance.integer_pgp_sym_field, expected)
 
     def test_null(self):
         """Assert `NULL` values are saved."""